moved helper function for streak to other file + made the progress items non-clickabe on congratulations page
